Finish homeworks
