+++ src/main/java/com/github/hcsp/annotation/CacheClassDecorator.java
@@ -20,6 +20,7 @@ public class CacheClassDecorator {
     // 注意，缓存的实现需要是线程安全的
     @SuppressWarnings("unchecked")
     public static <T> Class<T> decorate(Class<T> klass) {
+
         return (Class<T>) new ByteBuddy()
                 .subclass(klass)
                 .method(ElementMatchers.isAnnotatedWith(Cache.class))
@@ -29,94 +30,17 @@ public class CacheClassDecorator {
                 .getLoaded();
     }
 
-    private static class CacheKey {
-        private Object thisObject;
-        private String methodName;
-        private Object[] arguments;
+    public static void main(String[] args) throws Exception {
+        DataService dataService = decorate(DataService.class).getConstructor().newInstance();
 
-        CacheKey(Object thisObject, String methodName, Object[] arguments) {
-            this.thisObject = thisObject;
-            this.methodName = methodName;
-            this.arguments = arguments;
-        }
+        // 有缓存的查询：只有第一次执行了真正的查询操作，第二次从缓存中获取
+        System.out.println(dataService.queryData(1));
+        Thread.sleep(1 * 1000);
+        System.out.println(dataService.queryData(1));
 
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            CacheKey cacheKey = (CacheKey) o;
-            return Objects.equals(thisObject, cacheKey.thisObject) &&
-                    Objects.equals(methodName, cacheKey.methodName) &&
-                    Arrays.equals(arguments, cacheKey.arguments);
-        }
-
-        @Override
-        public int hashCode() {
-            int result = Objects.hash(thisObject, methodName);
-            result = 31 * result + Arrays.hashCode(arguments);
-            return result;
-        }
+        // 无缓存的查询：两次都执行了真正的查询操作
+        System.out.println(dataService.queryDataWithoutCache(1));
+        Thread.sleep(1 * 1000);
+        System.out.println(dataService.queryDataWithoutCache(1));
     }
-
-    private static class CacheValue {
-        private Object value;
-        private long time;
-
-        CacheValue(Object value, long time) {
-            this.value = value;
-            this.time = time;
-        }
-    }
-
-    public static class CacheAdvisor {
-        private static ConcurrentHashMap<CacheKey, CacheValue> cache = new ConcurrentHashMap<>();
-        @RuntimeType
-        public static Object cache(
-                @SuperCall Callable<Object> superCall,
-                @Origin Method method,
-                @This Object thisObject,
-                @AllArguments Object[] arguments) throws Exception {
-            CacheKey cacheKey = new CacheKey(thisObject, method.getName(), arguments);
-            final CacheValue resultExistingInCache = cache.get(cacheKey);
-
-            if (resultExistingInCache != null) {
-                if (cacheExpires(resultExistingInCache, method)) {
-                    return invokeRealMethodAndPutIntoCache(superCall, cacheKey);
-                } else {
-                    return resultExistingInCache.value;
-                }
-            } else {
-                return invokeRealMethodAndPutIntoCache(superCall, cacheKey);
-            }
-        }
-
-        private static Object invokeRealMethodAndPutIntoCache(@SuperCall Callable<Object> superCall, CacheKey cacheKey) throws Exception {
-            Object realMethodInvocationResult = superCall.call();
-            cache.put(cacheKey, new CacheValue(realMethodInvocationResult, System.currentTimeMillis()));
-            return realMethodInvocationResult;
-        }
-
-        private static boolean cacheExpires(CacheValue cacheValue, Method method) {
-            long time = cacheValue.time;
-            int cacheSeconds = method.getAnnotation(Cache.class).cacheSeconds();
-            return System.currentTimeMillis() - time > cacheSeconds * 1000;
-        }
-    }
-        public static void main(String[] args) throws Exception {
-            DataService dataService = decorate(DataService.class).getConstructor().newInstance();
-
-            // 有缓存的查询：只有第一次执行了真正的查询操作，第二次从缓存中获取
-            System.out.println(dataService.queryData(1));
-            Thread.sleep(1 * 1000);
-            System.out.println(dataService.queryData(1));
-
-            // 无缓存的查询：两次都执行了真正的查询操作
-            System.out.println(dataService.queryDataWithoutCache(1));
-            Thread.sleep(1 * 1000);
-            System.out.println(dataService.queryDataWithoutCache(1));
-        }
-    }
+}
